Bug 1833647 - Add support for feature detection to test262-update.py; r=anba
The Float16Array tests in test262 check for Float16Array on the fly instead of
using a feature flag. This adds support for checking the list of include files
for a specific include file, and adding the appropriate command line to the
existing shell option code if it is present.

Differential Revision: https://phabricator.services.mozilla.com/D205260

UltraBlame original commit: dd63daeb6c07563f32e9b61c6a62d10d2cdd3cf3