Bug 1893554 - Ensure Ion frames with rematerialized frames are invalidated. r=iain
When the debugger creates a `RematerializedFrame` for an Ion frame, we need to invalidate
the Ion frame to ensure the rematerialized frame is destroyed in `FinishBailoutToBaseline`
(or the exception handler).

The debugger code could fail with OOM before doing this invalidation and in this case we
left a rematerialized frame in the activation's map after returning from the Ion frame.

Differential Revision: https://phabricator.services.mozilla.com/D208774

UltraBlame original commit: a7ff2aad84a1e650dbf3c240814429f583ceb637