Bug 1893173 - Convert to TopSites into a singleton - r=dao
This doesn't necessarily prove it works as a singleton component since the
unit tests don't initialize it, but I figure it makes sense to convert the
unit test now before subsequent modifications.

The changes to the component code are due to the fact that previously,
the unit test would do one test after another with a new instance of a TopSitesFeed.
Since I'm re-using the instance, it needs to be reset. And since a
method could still be running when the next test runs, I wait until a
possible refresh is happening before continuing on to another test.

Differential Revision: https://phabricator.services.mozilla.com/D208996

UltraBlame original commit: 50c683aad9b449232a66468e22747397bc6a77c6