Bug 1894398 - validate treeherder group name length in taskcluster config. r=releng-reviewers,taskgraph-reviewers,jmaher,jcristau
Treeherder ingests tasks data and discards it if the Treeherder group name for
a task is longer than the limit of 100 characters in Treeherder - the task is
not shown in Treeherder. Enforce the limit in the Taskcluster schema to prevent
this.

Differential Revision: https://phabricator.services.mozilla.com/D209078

UltraBlame original commit: c6de63471506aadca84324d6ec8614589b17c798