Bug 1894882 - Add support for displaying long running jit tests r=dminor
This adds the --show-slow and --slow-test-threshold options to jit tests which
are already present for js tests.

This extends jstests to output this information via structured logging so it's
available in automation. It doesn't end up in the test log by default but it's
present in main_raw.log.

This also enables this reporting by default in automation.

Differential Revision: https://phabricator.services.mozilla.com/D209362

UltraBlame original commit: 1a9b33d9c18f7028f029f5dce855b4808ab6ae4c