Bug 1892398 - Don't hard-code that quad segments with a source render task are alpha blended. r=gw
Not immediately necessary but when cached quads or patterns that read a texture are implemented, There will be opaque segments taking render tasks as input. Currently the code is a bit error prone because some parts of the code advertize a certain behavior that is later overruled. This patch makes things more explicit and flexible.

Differential Revision: https://phabricator.services.mozilla.com/D208508

UltraBlame original commit: 986045c4cb9400087ab2f70842af82d31dcf602c