Bug 1895492: bump user agent of InetBgDL r=nalexander
This bump to the user agent allows us to distinguish between stub installers with and without this change, which enables us to keep both of them working after we switch to a new authenticode certificate in https://bugzilla.mozilla.org/show_bug.cgi?id=1889299

We do not have automatic builds for these; my plan is to use the child revisions of this (related to https://bugzilla.mozilla.org/show_bug.cgi?id=1771192) to build it on Try. From there, we'll sign it by hand (as we've done for plugins in the past), and check it into the tree in a separate revision that must be landed at the same time as the new authenticode certificate.

Differential Revision: https://phabricator.services.mozilla.com/D209700

UltraBlame original commit: 6af7fa4a42082b22efb0425b9d8c3cddd79fefe5