Bug 1887659 - Set mIsWaitingLongTapResult to false for mPrevActiveTouchBlock when a new touch-start is received. r=botond
mPrevActiveTouchBlock is set when a long-tap event happens and it waits for a
response from the contents whether contextmenu is preventDefault-ed or not. So
even if something went wrong in the content or somewhere else, resetting
mIsWaitingLongTapResult on every new touch-start event can avoid situtations
where the touch block gets stuck on waiting for the response.

Differential Revision: https://phabricator.services.mozilla.com/D209629

UltraBlame original commit: 6ce977896f7725edf062e6ebafe6dbab15b81aa8