Bug 1895398 - Change the about:profiling test assertion to only check the area it intends to check r=profiler-reviewers,florian
Previously this assertion was checking the whole preset values, including
interval, threads list and the enabled features. But if you look at the info
message above, it actually only intends to test a specific edge case of a
threads filter. So it's not really necessary to check the interval and the
features since that's not really important for this assertion.

I initially wanted to change this assertion because it caused multiple backouts
in the past when we changed the default values of some features. I think it's
better to not check the features here at all to make sure that we won't have
any more backouts in the future.

Differential Revision: https://phabricator.services.mozilla.com/D209612

UltraBlame original commit: 16c706919fcf22df1fae8499d4c05d7c88de5109