Bug 1887790 part 3: Expose aria-atomic via the UIA AriaProperties property. r=nlapre
UIA doesn't have a native property to expose aria-atomic, so we need to expose this via AriaProperties.
If it's not exposed, Narrator will assume true, so we must expose it even if it's not explicitly present.

As part of this, I removed the local-only implementation of AriaProperties.
While it was more complete, it's not useful to have this supported only for LocalAccessible.
In addition, it didn't expose implicit default values, which this atomic case proves we need in some cases.
We'll add more properties as needed.

Differential Revision: https://phabricator.services.mozilla.com/D209659

UltraBlame original commit: 2d77157cf946368065d0209bf775f5a5fdbf8ff4